requirements.txt
discord.py>=2
dot-env